feat(Cite): Every paragraph with starting --
Every paragraph with a startin -- is now interpreted as a citation and
rendered as a cite tag. If a quote contains a paragraph and a citation,
this is even more highlighted as a centered quotation.

This type of citation will be used in conjuction with more elements
later ... tables as caption or additionally with figures ...