improve(InlineTTS): PlayButtons labeled by content
The inline playback-buttons do now have labels, to identify the
area to be replayed, furthermore this extends the clickable area.

Next step ist to use `this.labels[0].innerText` to make the area
translateable by google to to refer to the text to read out loud.
Additonally markers need be be added to mark the text as not
translatable, if the language differs.