Adds hyper as a valid protocol for referencing
Previously only https:// and http:// were allowed for referencing. Now
it is also possible to refer to hyper:// which can be used within the
beaker browser.

A fallback or a check is currently not implemented, which might be
usefull if the user is within another browser, thus it might be
neccessary to use a proxy if one exists...