fix: Silent hidden TTS
The original TTS had been refactored earlier, such that comments are now
are printed to the DOM but they are hidden. This way, Google and other
translators can be used to also translate the spoken comments. However,
this was not done for hidden-TTS, they were still kept in an obsolete
string. There had been a couple of changes:

* normal comments will get hidden ones attached to them, to preserve
  order
* lonely hidden comments will be attached to the end of every document,
  this way the current TTS handling can find them and speak ...

[Closes: #54]