chore: Add parcel@1.12.3 for dev/watch mode
The new parcel bundler seem to work fine for building in production,
however it works poorly in dev or watch mode and does crash if there is
an elm error and does not update if there is an sass update. Thus, at
the moment the older version will be used in dev mode only.