fix(Image): Styling prevents the onclick Modal
If there is added custom styling to an image, then opening it as a modal
is prohibitted. The reason for this is, that custom styling causes a lot
of pain, such as scaling, the application of filters, ids, which are
used to manipulate an image, the application of image maps, which
require a click event. Also the application of image maps is only
possible, if the modal functionality is switched off ...