improve(Quiz): Vector now supports `- [[ ]]`
This allows to define quizzes and surveys similar to tasks with a
starting and OPTIONAL dash. The old syntax is still allowed...

Will result in an unordered list

- [[X]] True
- [[ ]] False

Use indentation for code view

  [[X]] True
  [[ ]] False

This will be treated differently by any other Markdown interpreter, but
in LiaScript these are identical quizzes.