refact(Persistent): Move unused funct. to Service
This functionality is currently not used anymore, but might be added in
the future again. The `Persistent` module takes care of all elements
that have a class-name persistent. These elements were previously stored
and if the section is visited again also restored from a local bag. But
this did not work well on all browsers.