feat: Add window.LIA.log
This global function can be overwritten to pass log information to an
external service. This can be used if an editor is used and the user
should also be able to get a glimpse onto the internal message-flow.
`window.LIA.debug` has to be set to true, in order to make debug
messages visible.