feat(SCORM2004): Add Connector
This connector allows to store quizzes, surveys, and tasks. For
simplicity, all of them are stored as SCORM interactions and the result
is stored within interactions.learner_response as long-fill-in type ...
This way, it is possible to store the state as JSON with newlines
included.

I do not know why SCORM is so restricive with variables and
model.order.foo ...