improve(Comments): Support for HTML comments
The parser now has a special rule for HTML comments, that do not contain
HTML parameters. These are now ignored and can be used for commenting.
But, comments with only words will still be interpreted as
HTML-parameters, since it is not possible to distinguish something like:

  <!-- autoplay data-set something not important -->

from a parameter set. This will be added to the HTML-block althoug it
does not make sense, or does it ;-)