improve(TTS): Waiting for translations
It might take some time until the translation from google is ready. This
change adds a timeout of two seconds to the first TTS output on every
slide. This way we can guarante that the google translation is ready and
also the DOM, which is then used to gather the text for speaking.