improve: Add persistent macro for DOM preservation
PyScript and other systems that also need to manipulate the DOM, require
persistency with their created nodes. LiaScript by default, creates for
each slide a new DOM. This way videos and sound and simulations stop,
when a new slide is loaded ... However, now it is possible to use the

  <!--
  persistent: true
  -->

macro. This will not destroy the slides anymore, but make them hidden.
The downside is, videos will not be stopped from playing anymore and
other heavy-wheight simulation will still be running in background.