fix: Import with relative resources
Previously relative imports from modules were not possible, but this
changed with the latest release. This introduced a new bug, which was
previously not possible:

"A module import can load relative URLs (js, css)"

This was not handled and the wrong base-path was set while loading. Now
the URL of the import module is used as the relative base for all
subsequent loads.

fixes: 108