improve: oEmbed-provider extracted from website
As stated by the spec in:

https://oembed.com

it is also possible to extract the oEmbed provider from the website
itself, by parsing the header for:

  <link rel="alternate" type="application/json+oembed"
  href="http://flickr.com/services/oembed?url=...."
  title="Bacon..." />

This is done now, after the provider list is checked. According to the
spec it should be done in the opposite way, but there are websites like
youtube that will block a previous request. Additionally, this may
reduce networktraffic.