refactor: oEmbed provider grabber
The python tool for generating the endpoint.ts now performs a check to
all URLs providen from:

https://oembed.com/providers.json

if the provider seems to be dead, then it not included into the
endpoints list and a note is added the end of the generated README.md
with provider name, url, and error messages.

Additionally preview-images are extracted during the grabbing and these
images are added to the generated Markdown-entry as well, if possible.