feat: Add macros to formula
Macros to formulas can no be defined globally with the formula macro

  <!--
  formula: Theta   \mathrm{\zeta}(#1)
  formula: \beta   \Theta{B}
  -->

These macros are then passed to every formula. If formulas are defined
with \def or \gdef, etc. within a formula, then only these macros are
used an no external ones are passed.

Additionally, formulas from other imports are imported and used as well.
This way it is possible to include reusable formula collections.