improve(Macro): Parameter with more than 3 backticks
The macro parameters formats had the following format

- `single,line,parameter,with,commas`
- ``` multiline `parameters` ```

The 3 backticks were hard-coded, hence it was not possible to pass
markdown-blocks with code-blocks. Now it is possible to define multiline
parameters with more than 3 backticks, allowing for more complex
parameters. Basically it works like an ordinary code-block now, which
can contain other code-blocks with less backticks.